# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API service for the District Core Crime Management System",
        "title": "District Core Crime Management System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/evidence/image": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload an image as evidence for a case",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evidence"
                ],
                "summary": "Upload image evidence",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Case ID",
                        "name": "caseId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional remarks about the evidence",
                        "name": "remarks",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Image file",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Evidence"
                        }
                    },
                    "400": {
                        "description": "Invalid request or not an image",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/evidence/text": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add textual evidence to a case",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evidence"
                ],
                "summary": "Create text evidence",
                "parameters": [
                    {
                        "description": "Text evidence details",
                        "name": "evidence",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_dto.CreateTextEvidenceDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Evidence"
                        }
                    },
                    "400": {
                        "description": "Invalid evidence data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/evidence/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve details of a specific evidence item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evidence"
                ],
                "summary": "Get evidence details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Evidence ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Evidence"
                        }
                    },
                    "400": {
                        "description": "Invalid evidence ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Evidence not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update evidence remarks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evidence"
                ],
                "summary": "Update evidence",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Evidence ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated evidence details",
                        "name": "evidence",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_dto.UpdateEvidenceDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Evidence"
                        }
                    },
                    "400": {
                        "description": "Invalid evidence ID or data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mark evidence as deleted (soft delete)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evidence"
                ],
                "summary": "Soft delete evidence",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Evidence ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid evidence ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/evidence/{id}/audit": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve audit logs for a specific evidence item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evidence",
                    "audit"
                ],
                "summary": "Get evidence audit logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Evidence ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.AuditLog"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid evidence ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/evidence/{id}/image": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Stream an evidence image file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "image/*"
                ],
                "tags": [
                    "evidence"
                ],
                "summary": "Get evidence image",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Evidence ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid evidence ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error or not an image",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/evidence/{id}/permanent": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permanently delete evidence (requires confirmation)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evidence"
                ],
                "summary": "Permanently delete evidence",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Evidence ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Deletion confirmation",
                        "name": "confirmation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_dto.DeleteConfirmationDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid evidence ID or missing confirmation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_m7medVision_crime-management-system_internal_dto.CreateTextEvidenceDTO": {
            "type": "object",
            "required": [
                "caseId",
                "content"
            ],
            "properties": {
                "caseId": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                }
            }
        },
        "github_com_m7medVision_crime-management-system_internal_dto.DeleteConfirmationDTO": {
            "type": "object",
            "required": [
                "confirmation"
            ],
            "properties": {
                "confirmation": {
                    "type": "string"
                }
            }
        },
...
        "github_com_m7medVision_crime-management-system_internal_dto.UpdateEvidenceDTO": {
            "type": "object",
            "properties": {
                "remarks": {
                    "type": "string"
                }
            }
        },
...
        "github_com_m7medVision_crime-management-system_internal_model.AuditLog": {
            "type": "object",
            "properties": {
                "action": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.ActionType"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "entityID": {
                    "type": "integer"
                },
                "entityType": {
                    "description": "\"evidence\", \"case\", \"user\", etc.",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ipaddress": {
                    "type": "string"
                },
                "newValue": {
                    "type": "string"
                },
                "oldValue": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_m7medVision_crime-management-system_internal_model.Evidence": {
            "type": "object",
            "properties": {
                "addedBy": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.User"
                },
                "addedByID": {
                    "type": "integer"
                },
                "case": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Case"
                },
                "caseID": {
                    "type": "integer"
                },
                "content": {
                    "description": "For text evidence",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "imagePath": {
                    "description": "For image evidence",
                    "type": "string"
                },
                "isDeleted": {
                    "description": "For soft delete",
                    "type": "boolean"
                },
                "remarks": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.EvidenceType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_m7medVision_crime-management-system_internal_model.EvidenceType": {
            "type": "string",
            "enum": [
                "text",
                "image"
            ],
            "x-enum-varnames": [
                "EvidenceTypeText",
                "EvidenceTypeImage"
            ]
        },
...
        "github_com_m7medVision_crime-management-system_internal_model.User": {
            "type": "object",
            "properties": {
                "clearanceLevel": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.ClearanceLevel"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "lastLogin": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Role"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}